Rename data to locations
